+++ CHANGELOG.md
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed unnecessary imports and simplify code
 - Added performance tests for `StringSegmentLinkedList`
 
+## [0.4.2] / 2025-11-11
+### 🐛 Fixes
+- Add missing `Parse` and `TryParse` methods for `NonNegativeInteger` and `PositiveInteger` types.
+- Add mising `PositiveInteger.ToString()` to better display positive integer values in logs.
+
+
 ## [0.4.1] / 2025-08-12
 ### 🐛 Fixes
 - Revert usage of weak references to optimize `StringSegmentLinkedList` memory consumption.
@@ -127,7 +133,8 @@ which is now redundant
 ## [0.1.0] / 2023-01-29
 - Initial release
 
-[Unreleased]: https://github.com/candoumbe/Candoumbe.Types/compare/0.4.1...HEAD
+[Unreleased]: https://github.com/candoumbe/Candoumbe.Types/compare/0.4.2...HEAD
+[0.4.2]: https://github.com/candoumbe/Candoumbe.Types/compare/0.4.1...0.4.2
 [0.4.1]: https://github.com/candoumbe/Candoumbe.Types/compare/0.4.0...0.4.1
 [0.4.0]: https://github.com/candoumbe/Candoumbe.Types/compare/0.3.1...0.4.0
 [0.3.1]: https://github.com/candoumbe/Candoumbe.Types/compare/0.3.0...0.3.1